split `DragPreview` and `ViewSyncer` from `InventoryView`
