+++ Packages/com.fullmetalbagel.dope-inventory-ugui/Runtime/UI/InventoryDragPreviewController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 822efea37163f4dbf8283fe8696d268b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
