+++ Packages/com.fullmetalbagel.dope-inventory-ugui/Runtime/UI/InventoryViewSyncer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b9503534cec6044b2a339a4c2cf2cd2b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
